requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
apache-airflow==2.10.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.1
apache-airflow-providers-common-sql==1.17.1
apache-airflow-providers-fab==1.4.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-google==10.23.0
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apache-beam==2.58.1
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
blinker==1.8.2
cachelib==0.9.0
cachetools==5.4.0
cattrs==24.1.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.8.2
ConfigUpdater==3.2
connexion==2.14.2
crcmod==1.7
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.1
db-dtypes==1.3.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastavro==1.9.5
fasteners==0.19
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2024.9.0.post1
google-ads==25.0.0
google-analytics-admin==0.23.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-apitools==0.5.31
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.64.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-automl==2.13.5
google-cloud-batch==0.17.28
google-cloud-bigquery==3.25.0
google-cloud-bigquery-datatransfer==3.15.7
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.26.0
google-cloud-build==3.25.0
google-cloud-compute==1.19.2
google-cloud-container==2.51.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.20.1
google-cloud-dataflow-client==0.8.12
google-cloud-dataform==0.5.11
google-cloud-dataplex==2.2.2
google-cloud-dataproc==5.13.0
google-cloud-dataproc-metastore==1.15.5
google-cloud-datastore==2.20.1
google-cloud-dlp==3.22.0
google-cloud-kms==3.0.0
google-cloud-language==2.14.0
google-cloud-logging==3.11.2
google-cloud-memcache==1.9.5
google-cloud-monitoring==2.22.2
google-cloud-orchestration-airflow==1.14.0
google-cloud-os-login==2.14.6
google-cloud-pubsub==2.23.0
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.12
google-cloud-redis==2.15.5
google-cloud-resource-manager==1.12.5
google-cloud-run==0.10.8
google-cloud-secret-manager==2.20.2
google-cloud-spanner==3.48.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.12.0
google-cloud-tasks==2.16.5
google-cloud-texttospeech==2.17.2
google-cloud-translate==3.16.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-cloud-workflows==1.14.5
google-crc32c==1.5.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.65.4
grpcio-gcp==0.2.2
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.7
immutabledict==4.2.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Js2Py==0.74
json-merge-patch==0.2
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.16.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.7.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandas-gbq==0.23.2
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
py4j==0.10.9.5
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydata-google-auth==1.8.2
pydot==1.4.2
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.9.0
pymongo==4.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyspark==3.3.2
pytest==8.3.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.9
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.13.1
zipp==3.20.2
zstandard==0.23.0